+++ ai/src/test/kotlin/ktx/ai/Cat.kt
@@ -1,0 +1,3 @@
+package ktx.ai
+
+class Cat