+++ ai/src/test/kotlin/ktx/ai/EatTask.kt
@@ -1,0 +1,19 @@
+package ktx.ai
+
+import com.badlogic.gdx.ai.btree.LeafTask
+import com.badlogic.gdx.ai.btree.Task
+import com.badlogic.gdx.ai.btree.annotation.TaskAttribute
+
+class EatTask(
+  @JvmField
+  @TaskAttribute(required = true)
+  var hunger: Int,
+) : LeafTask<Cat>() {
+  override fun copyTo(task: Task<Cat>): Task<Cat> {
+    return this
+  }
+
+  override fun execute(): Status {
+    return Status.SUCCEEDED
+  }
+}